+++ enum/language.enum.ts
@@ -1,0 +1,5 @@
+export enum Language {
+  IT = "IT",
+  EN = "EN",
+  JP = "JP",
+}
