+++ enum/tagCondition.enum.ts
@@ -1,0 +1,7 @@
+export enum TagCondition {
+  NM = "NM",
+  EX = "EX",
+  GD = "GD",
+  PL = "PL",
+  PO = "PO",
+}
