+++ interface/checkoutForm.interface.ts
@@ -1,0 +1,9 @@
+export interface CheckoutForm {
+  name: string;
+  surname: string;
+  email: string;
+  cap: string;
+  city: string;
+  streetAndHouseNumber: string;
+  iWantTheInvoice: boolean;
+}
