+++ interface/highlightResult.interface.ts
@@ -1,0 +1,13 @@
+export interface HighlightResult {
+  name: {
+    value: string;
+    matchLevel: string;
+    fullyHighlighted: boolean;
+    matchedWords: string[];
+  };
+  tags: Array<{
+    value: string;
+    matchLevel: string;
+    matchedWords: string[];
+  }>;
+}
