+++ interface/linkRoute.interface.ts
@@ -1,0 +1,4 @@
+export interface LinkRoute {
+  link: string;
+  route: string;
+}
