+++ interface/product.interface.ts
@@ -1,0 +1,6 @@
+export interface Product {
+  nameProduct: string;
+  price: number;
+  codeProduct: string;
+  id: string;
+}
