+++ interface/searchProductResult.interface.ts
@@ -1,0 +1,26 @@
+import type { Hit } from "./hit.interface";
+
+export interface SearchProductResult {
+  hits: Hit[];
+  nbHits: number;
+  page: number;
+  nbPages: number;
+  hitsPerPage: number;
+  exhaustiveNbHits: boolean;
+  exhaustiveTypo: boolean;
+  exhaustive: {
+    nbHits: boolean;
+    typo: boolean;
+  };
+  query: string;
+  params: string;
+  renderingContent: Record<string, unknown>;
+  processingTimeMS: number;
+  processingTimingsMS: {
+    _request: {
+      roundTrip: number;
+    };
+    total: number;
+  };
+  serverTimeMS: number;
+}
