+++ types/headerPropsType.type.ts
@@ -1,0 +1,3 @@
+export type HeaderProps = {
+  cartCount: number;
+};
