+++ types/newsBanner.type.ts
@@ -1,0 +1,4 @@
+export type NewsBanner = {
+  label: string;
+  url?: string;
+};
