+++ types/paymentMethod.type.ts
@@ -1,0 +1,5 @@
+export type PaymentMethod = {
+  src: string;
+  alt: string;
+  link: string;
+};
