+++ types/productCard.type.ts
@@ -1,0 +1,11 @@
+export type ProductCard = {
+  colorScheme?: string;
+  id?: string;
+  tcg?: string;
+  category?: string;
+  productName: string;
+  code: string;
+  expansion: string;
+  price: string;
+  imageUrl: string;
+};
