+++ types/productType.type.ts
@@ -1,0 +1,10 @@
+export type ProductType = {
+  id: string;
+  productName: string;
+  code: string;
+  expansion: string;
+  price: string;
+  imageUrl: string;
+  tcg: string;
+  category: string;
+};
