+++ types/socialLink.type.ts
@@ -1,0 +1,4 @@
+export type SocialLinks = {
+  img: string;
+  url: string;
+};
