+++ types/tagCode.type.ts
@@ -1,0 +1,4 @@
+import type { Language } from "~/enum/language.enum";
+import type { TagCondition } from "~/enum/tagCondition.enum";
+
+export type TagCode = TagCondition | Language;
