+++ types/tagStructure.type.ts
@@ -1,0 +1,13 @@
+import type { Language } from "~/enum/language.enum";
+import type { TagCondition } from "~/enum/tagCondition.enum";
+
+export type TagStructure = {
+  language: Language.IT | Language.EN | Language.JP;
+  conditions: (
+    | TagCondition.NM
+    | TagCondition.EX
+    | TagCondition.GD
+    | TagCondition.PL
+    | TagCondition.PO
+  )[];
+};
